+++ tuf/requests_fetcher.py
@@ -12,10 +12,10 @@
   See LICENSE-MIT OR LICENSE for licensing information.
 
 <Purpose>
-  Provides an interface for network IO abstraction.
+  Provides an implementation of FetcherInterface using the requests HTTP
+  library.
 """
 
-import abc
 import requests
 import six
 import logging
@@ -24,46 +24,12 @@ import time
 import urllib3.exceptions
 import tuf.exceptions
 import tuf.settings
+import tuf.client.fetcher
 
 logger = logging.getLogger(__name__)
 
 
-class FetcherInterface():
-  """
-  <Purpose>
-  Defines an interface for abstract network download which can be implemented
-  for a variety of network libraries and configurations.
-  """
-  __metaclass__ = abc.ABCMeta
-
-
-  @abc.abstractmethod
-  def fetch(self, url, required_length):
-    """
-    <Purpose>
-      Fetches the contents of HTTP/HTTPS url from a remote server up to
-      required_length and returns a bytes iterator.
-
-    <Arguments>
-      url:
-        A URL string that represents the location of the file.
-
-      required_length:
-        An integer value representing the length of the file in bytes.
-
-    <Exceptions>
-      tuf.exceptions.SlowRetrievalError, if a timeout occurs while receiving
-      data from a server
-
-      tuf.exceptions.FetcherHTTPError, if an HTTP error code was received
-    <Returns>
-      A bytes iterator
-    """
-    raise NotImplementedError # pragma: no cover
-
-
-
-class RequestsFetcher(FetcherInterface):
+class RequestsFetcher(tuf.client.fetcher.FetcherInterface):
   """
   <Purpose>
     A concrete implementation of FetcherInterface based on the Requests
